textproc/denote.el: new port
Denote aims to be a simple-to-use, focused-in-scope, and effective
note-taking and file-naming tool for Emacs.

Denote is based on the idea that files should follow a predictable and
descriptive file-naming scheme. The file name must offer a clear
indication of what the contents are about, without reference to any
other metadata. Denote basically streamlines the creation of such
files or file names while providing facilities to link between them
(where those files are editable).

PR:		276317
Approved by:	submitter is maintainer